% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_budyko.R
\name{compute_budyko}
\alias{compute_budyko}
\title{compute_budyko}
\usage{
compute_budyko(P, PET, w = 2.63)
}
\arguments{
\item{P}{precipitation}

\item{PET}{potential evapotranspiration}

\item{w}{adjustment parameter (default 2.63)}
}
\value{
Evaporative index (ratio)
}
\description{
THis evaporative index (ratio of AET/P) using Budyko
}
\author{
Naomi
}
