% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_budyko2.R
\name{compute_budyko2}
\alias{compute_budyko2}
\title{compute_budyko2}
\usage{
compute_budyko2(P, PET, w = 2.63)
}
\arguments{
\item{P}{precipitation}

\item{PET}{potential evapotranspiration}

\item{w}{adjustment parameter (default 2.63)}
}
\value{
Evaporative index (ratio), Q (streamflow), AET (actual evapotranspration)
}
\description{
THis evaporative index (ratio of AET/P) using Budyko
}
\author{
Naomi
}
