% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mission.R
\docType{data}
\name{mission}
\alias{mission}
\title{Climate Data from Mission Creek in California}
\format{A dataframe with 11 columns and 22617 rows
\itemize{
\item date Month/day/year
 \item tmax maximum daily temperature C
\item tmin minimum daily temperature C
\item rain rainfall in mm
\item year
\item month
\item day
\item wy water year
\item cfs streamflow in cfs (volume)
\item wyd day of water year
\item yr day of calendar year
\item Q streamflow in mm (area averaged)
}}
\usage{
mission
}
\description{
Data from Rocky Nook meterologic station and USGS Gage at Mission Creek Outflow (21km2)
}
\keyword{datasets}
