% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_montieth.R
\name{penman_montieth}
\alias{penman_montieth}
\title{penman_montieth}
\usage{
penman_montieth(Tair, vpd, Rnet, gs, ga, CP = 1010, Pair = 101325,
  dayl = 24)
}
\arguments{
\item{Tair}{(deg C) air temperature}

\item{vpd}{(Pa)    vapour pressure deficit}

\item{Rnet}{(j/m2/day)      net radiation}

\item{gs}{(mm/2)  surface conductance}

\item{ga}{(mm/s)  aerodynamic conductance}

\item{CP}{=      1010.0 (J/kg*K) specific heat of air}

\item{Pair}{air pressure (default 101325 (Pa)}

\item{dayl}{number of hours in the day (default 24)}

\item{newp}{sensitivity analysis default=1}
}
\value{
Evapotranspiration (mm/day)
}
\description{
THis function computer evapotranspiration based on radiation, conductance etc
}
\author{
Naomi
}
