% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageclim.R
\docType{data}
\name{sageclim}
\alias{sageclim}
\title{Climate Data from Sagehen Creek}
\format{A dataframe with 8 columns and 27274 row
\itemize{
\item date Month/day/year
\item tmin minimum daily temperature C
\item tmax maximum daily temperature C
\item rain rainfall in mm
\item year
\item month
\item day
\item wy water year
}}
\usage{
sageclim
}
\description{
Data from meterologic station
}
\keyword{datasets}
