% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagecomb.R
\docType{data}
\name{sagecomb}
\alias{sagecomb}
\title{Climate Data from Sagehen Creek}
\format{A dataframe with 11 columns and 22617 rows
\itemize{
\item date Month/day/year
\item year
\item month
\item day
\item wy water year
\item tmax maximum daily temperature C
\item tmin minimum daily temperature C
\item rain rainfall in mm
\item Q streamflow in mm
\item sTmax maximum daily stream temperature C
\item sTmin minimum daily stream temperature C
}}
\usage{
sagecomb
}
\description{
Data from meterologic station
}
\keyword{datasets}
