% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbddata.R
\docType{data}
\name{sbdata}
\alias{sbdata}
\title{Climate Data from Santa Barbara}
\format{A dataframe with 11 columns and 22617 rows
\itemize{
\item date Month/day/year
\item year
\item month
\item day
\item wy water year
\item tavg average daily temperature C
\item precip rainfall in mm
\item vpd vapor pressure deficit in Pa
\item rnet net shortwave radiation (j/m2/day)
}}
\usage{
sbdata
}
\description{
Data from meterologic station
}
\keyword{datasets}
