% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.R
\docType{data}
\name{sun}
\alias{sun}
\title{Direct solar radiation and photosynthesis at top of canopy}
\format{An object of class \code{data.frame} with 1461 rows and 10 columns.}
\usage{
sun
}
\description{
Data from Sagehen simulation model (estimates based on latitude, day of year and solar geometry
and estimates of atmospheric transmissivity)
}
\details{
@format A dataframe with 5 columns and 1461 row
\itemize{
\item Kdown_direct downwelling solar radiation (J/m2/day)
\item year
\item month
\item day
\item wy water year
\item psn kgC/m2/day
}
}
\keyword{datasets}
